LPS-56502 Move 7.0.0 journal upgrade processes to module
