+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_7_0_0.java
@@ -30,9 +30,6 @@ import com.liferay.portal.upgrade.v7_0_0.UpgradeEmailAddress;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeEmailNotificationPreferences;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeExpando;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeGroup;
-import com.liferay.portal.upgrade.v7_0_0.UpgradeJournal;
-import com.liferay.portal.upgrade.v7_0_0.UpgradeJournalArticleType;
-import com.liferay.portal.upgrade.v7_0_0.UpgradeJournalDisplayPreferences;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeListType;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeMembershipRequest;
 import com.liferay.portal.upgrade.v7_0_0.UpgradeMessageBoards;
@@ -77,9 +74,6 @@ public class UpgradeProcess_7_0_0 extends UpgradeProcess {
 		upgrade(UpgradeEmailNotificationPreferences.class);
 		upgrade(UpgradeExpando.class);
 		upgrade(UpgradeGroup.class);
-		upgrade(UpgradeJournal.class);
-		upgrade(UpgradeJournalDisplayPreferences.class);
-		upgrade(UpgradeJournalArticleType.class);
 		upgrade(UpgradeListType.class);
 		upgrade(UpgradeMembershipRequest.class);
 		upgrade(UpgradeMessageBoards.class);
